Move rxtx_ring_init() to rxtx_ring.c.
