+++ rxtx_ring.h
@@ -11,6 +11,7 @@
 
 #include "rxtx.h" // for rxtx_ring
 
+int rxtx_ring_init(struct rxtx_ring *p, struct rxtx_desc *rtd, int ring_idx, char *errbuf);
 int rxtx_ring_destroy(struct rxtx_ring *p);
 int rxtx_ring_mark_packets_in_buffer_as_unreliable(struct rxtx_ring *p);
 int rxtx_ring_savefile_open(struct rxtx_ring *p, const char *template);
